+++ .gitignore
@@ -9,3 +9,5 @@ lang/*
 
 _data/i18n/*.yml
 !_data/i18n/en.yml
+node_modules/
+js/build
