+++ _layouts/default.html
@@ -47,25 +47,5 @@
     {% for script in page.scripts %}
       <script type="text/javascript" src="{{script}}"></script>
     {% endfor %}
-
-    {% if jekyll.environment == "production" %}
-      <script>
-        // Takes over language dropdown to set cookie
-        var langMenu = document.getElementById('dropdownNavLanguageMenu');
-        var langMenuItems = langMenu.getElementsByClassName('dropdown-item');
-
-        for (var i = 0; i < langMenuItems.length; i++) {
-          var langMenuItem = langMenuItems[i];
-
-          langMenuItem.addEventListener('click', function(event) {
-            var target = event.currentTarget;
-            var lang = target.dataset.lang;
-            var date = new Date();
-            date.setFullYear(date.getFullYear() + 1);
-            document.cookie = 'nf_lang=' + lang + ';path=/;expires=' + date.toUTCString();
-          });
-        }
-      </script>
-    {% endif %}
   </body>
 </html>
