+++ _layouts/pages/docs.html
@@ -1,8 +1,7 @@
 ---
 layout: page
 scripts:
-  - https://cdn.jsdelivr.net/docsearch.js/1/docsearch.min.js
-  - /js/docsearch.js
+  - "/js/build/documentation.js"
 ---
 
 {% include vars.html %}
