+++ _layouts/pages/install-ci.html
@@ -1,7 +1,7 @@
 ---
 layout: guide
 scripts:
-  - "/js/deeplink.js"
+  - "/js/build/install.js"
 ---
 
 {% include vars.html %}
