+++ _layouts/pages/install.html
@@ -1,8 +1,7 @@
 ---
 layout: guide
 scripts:
-  - "/js/install.js"
-  - "/js/deeplink.js"
+  - "/js/build/install.js"
 ---
 
 {% include vars.html %}
