+++ _layouts/pages/nightly.html
@@ -1,7 +1,7 @@
 ---
 layout: guide
 scripts:
-  - "/js/nightly.js"
+  - "/js/build/nightly.js"
 ---
 
 {% include vars.html %}
