+++ js/src/common.js
@@ -1,0 +1,5 @@
+import {fillLanguageDropdown} from './lib/production';
+
+if (process.env.NODE_ENV === 'production') {
+  production();
+}
