+++ js/src/documentation.js
@@ -1,3 +1,5 @@
+import docsearch from 'docsearch.js';
+
 docsearch({
   apiKey: '3949f721e5d8ca1de8928152ff745b28',
   indexName: 'yarnpkg',
