+++ js/src/install.js
@@ -1,0 +1,4 @@
+import {handleTabs, handleOS} from './lib/deeplink';
+
+handleTabs();
+handleOS();
