+++ search.html
@@ -4,9 +4,7 @@ layout: page
 stylesheets:
   - "https://cdn.jsdelivr.net/instantsearch.js/1.10.2/instantsearch.min.css"
 scripts:
-  - "https://cdn.jsdelivr.net/momentjs/2.17.1/moment.min.js"
-  - "https://cdn.jsdelivr.net/instantsearch.js/1.10.2/instantsearch-preact.js"
-  - "/js/search.js"
+  - "/js/build/search.js"
 title: Package Search
 featured_packages:
   -
