+++ webpack.config.js
@@ -1,0 +1,34 @@
+const webpack = require('webpack');
+
+module.exports = {
+  entry: {
+    common: './js/src/common.js',
+    documentation: './js/src/documentation.js',
+    install: './js/src/install.js',
+    nightly: './js/src/nightly.js',
+    search: './js/src/search.js',
+  },
+  output: {
+    path: './js/build',
+    filename: '[name].js',
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        use: [
+          'babel-loader',
+        ],
+        exclude: /node_modules/
+      }
+    ]
+  },
+  plugins:[
+    new webpack.DefinePlugin({
+      'process.env': {
+        NODE_ENV: JSON.stringify('production')
+      }
+    }),
+    new webpack.optimize.UglifyJsPlugin()
+  ]
+}
